Move window width size class to core UI layer
