+++ app/src/main/kotlin/com/d4rk/android/apps/apptoolkit/app/apps/common/screens/loading/HomeLoadingScreen.kt
@@ -12,7 +12,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import com.d4rk.android.libs.apptoolkit.core.utils.constants.ui.SizeConstants
 import com.d4rk.android.libs.apptoolkit.core.utils.platform.WindowItemFit
-import com.d4rk.android.libs.apptoolkit.core.utils.window.AppWindowWidthSizeClass
+import com.d4rk.android.libs.apptoolkit.core.ui.window.AppWindowWidthSizeClass
 
 @Composable
 fun HomeLoadingScreen(
